requirements.txt
PyQt5==5.15.9
pyttsx3==2.90
python-dotenv==1.0.0
google-generativeai==0.3.1